chore: replace sassc with dartsass
`sassc` uses LibSass, the C implementation of SASS. This library is
deprecated, and it can't parse certain SASS features, e.g. this:

```
color: hsla(var(--some-variable), 1);
```

These features are used in current versions of the gem `graphiql-rails`
though, so we have to replace sassc with dartsass in the `honeypot`
repository.

Now using `dartsass` for the main `honeypot` app, and including this
repository as a gem using `sassc` leads to errors like this:

```
TSort::Cyclic: topological sort failed:
[
	#<Rails::Initializable::Initializer:...
    @name=:setup_compression,
    @context=#<SassC::Rails::Railtie>,
    @options={:group=>:all, :after=>:setup_sass},
    @block=#<Proc:... /.../gems/sassc-rails-2.1.2/lib/sassc/rails/railtie.rb:72>
  >,
  #<Rails::Initializable::Initializer:...
    @name=:setup_sass,
    @context=#<SassC::Rails::Railtie>,
    @options={:group=>:all, :after=>:setup_compression},
    @block=#<Proc:... /usr/.../gems/dartsass-sprockets-3.1.0/lib/sassc/rails/railtie.rb:36>
  >,
  #<Rails::Initializable::Initializer:...
    @name=:setup_compression,
    @context=#<SassC::Rails::Railtie>,
    @options={:group=>:all, :after=>:setup_sass},
    @block=#<Proc:... /.../gems/dartsass-sprockets-3.1.0/lib/sassc/rails/railtie.rb:74>
  >
] (TSort::Cyclic)
```

Hence we have to use `dartsass` in this repository as well.

For details please see https://github.com/honeypotio/honeypot/pull/6709